+++ app/src/main/kotlin/com/wire/android/ui/home/conversationslist/MentionScreen.kt
@@ -4,13 +4,13 @@ import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.LazyListState
-import androidx.compose.foundation.lazy.rememberLazyListState
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import com.wire.android.R
 import com.wire.android.ui.common.dimensions
+import com.wire.android.ui.common.extension.rememberLazyListState
 import com.wire.android.ui.home.conversationslist.model.ConversationType
 import com.wire.android.ui.home.conversationslist.model.ConversationUnreadMention
 import com.wire.android.ui.home.conversationslist.model.EventType
@@ -24,8 +24,9 @@ fun MentionScreen(
     onEditConversationItem: (ConversationType) -> Unit,
     onScrollPositionChanged: (Int) -> Unit = {}
 ) {
-    val lazyListState = rememberLazyListState()
-    onScrollPositionChanged(lazyListState.firstVisibleItemIndex)
+    val lazyListState = rememberLazyListState { firstVisibleItemIndex ->
+        onScrollPositionChanged(firstVisibleItemIndex)
+    }
 
     MentionContent(
         lazyListState = lazyListState,
